requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==2.4.1
attrs==25.3.0
bitarray==3.3.1
blinker==1.9.0
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
cytoolz==1.0.1
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.12
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
instructor==1.7.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.0.24
langgraph-sdk==0.1.61
langsmith==0.3.30
lark==1.2.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
multidict==6.4.3
narwhals==1.34.1
numpy==2.2.4
openai==1.73.0
openpyxl==3.1.5
orjson==3.10.16
ormsgpack==1.9.1
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.2.1
pluggy==1.5.0
portia-sdk-python==0.1.14
propcache==0.3.1
protobuf==5.29.4
py-solc-x==2.0.3
pyarrow==19.0.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solc==0.0.0a0
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
streamlit==1.44.1
streamlit-chat==0.1.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vyper==0.4.1
web3==7.10.0
websockets==15.0.1
wheel==0.45.1
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
